Do not ask for password if not needed.
In certain circumstances we do not need to ask for a password because we
already have previously authenticated to sudo and can just reuse the
credentials.

Since there is no way to check if a sudo timestamp has timed out we run
a simple command and check the result.

Using an existing sudo timestamp from within nvim seems to work only if
the timestamp_type is set to 'global' in the sudo configuration file.

The assesment of the risk of setting the security relevant setting
timestamp_type to 'global' needs to be done by the user.