refactor!: change suda#system to a list-based api (#74)
* refactor!: change suda#system to a list-based api

This eliminates the need for shell-escaping, and removes possibility of accidental shell-injection.

This is technically a breaking change, since suda#system is globally exposed; however this is *not* a documented public api, so the chance of breakage is minimal.

Note that this *changes the behavior of g:suda#executeable! It is no longer shell-processed, but rather treated as a path/name to an executable and processed by the os in that way.
Should we split suda#executable on spaces to preserve the old behavior? Should we allow suda#executable to (optionally) be a list to allow passing arguments?

* refactor: implement suda#systemlist

This replaces the 3 argument suda#system implementation and follows the vimscript implementation.

* fix: #system passes empty list to systemlist

use call() instead to ensure that it passes the right amount of args