Remove `nobackup` option (#86)
Close #85

The `backup` (and `nobackup`) option is global and not local to the
buffer so this change affects all buffers. This is not the intended.

Note that it seems the `backup` option does not have any effect while
suda write the file through `BufWriteCmd`. So this simple change should
not affect the security of the plugin.